use locale : group pattern to speed up extraction
